Refactored Ivar modules yet again
This time joined IvarReadable and IvarWriteable into the same namespace
(Ivar) and using submodules Readable and Writeable for both.

I hope this is the final iteration of this files but one never knows
hah.
