+++ lib/aye_commander.rb
@@ -1,7 +1,6 @@
 require 'aye_commander/abortable'
 require 'aye_commander/inspectable'
-require 'aye_commander/ivar_readable'
-require 'aye_commander/ivar_writeable'
+require 'aye_commander/ivar'
 require 'aye_commander/limitable'
 require 'aye_commander/status'
 require 'aye_commander/errors'
